furniture_api/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==45.0.7
dj-rest-auth==7.0.1
Django==5.1.7
django-allauth==65.5.0
django-filter==25.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
packaging==25.0
psycopg==3.2.6
psycopg-binary==3.2.6
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.9.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.13.0
uritemplate==4.2.0
urllib3==2.3.0
whitenoise==6.9.0
